auth/requirements.txt
Flask==2.0.3
six==1.16.0
google-auth==2.6.2
requests==2.27.1